remvoes int oveflow https://stackoverflow.com/a/51888813/2602477

+++ src/utils.h
@@ -2,7 +2,7 @@
 #define utils_H
 #include "Rcpp.h"
 
-int wu_calc(int d, int d_a, int d_b, int x_a, int x_b);
+double wu_calc(int d, int d_a, int d_b, int x_a, int x_b);
 int count_permutations(arma::vec number);
 int is_power_of_two(unsigned int x);
